Opt-in support for OSC52 terminal clipboard access
Usernames and passwords can be copied to the clipboard using CTRL-U and
CTRL-P.  Tupm supports several options for performing clipboard copy,
such as calling the external xsel program.  This change supports using
the xterm-style OSC52 escape sequence for clipboard copy.  Because these
escape sequences may cause interference on terminals that don't support
them, this support is currently only enabled when the OSC52 environment
variable is set.

The motivation for this change is to support clipboard copy while using
the Chrome OS terminal.

For details on OSC52, see:
http://invisible-island.net/xterm/ctlseqs/ctlseqs.html