Fix: Don't exit in library code.
The upm create (as opposed to tupm) is a library which may potentially
be used by other programs in a variety of environments. (For example,
GUI programs, web application backends, etc.) Thus, it should not exit
or print to standard error, but simply return an error to the caller
when such a condition occurs.